more refactoring of 1.2 and 2.0 handlers
