+++ packages/main/src/handlers/fdc3/1.2/open.ts
@@ -1,12 +1,8 @@
 import { getRuntime } from '/@/index';
-import { View } from '/@/view';
 import { OpenError } from 'fdc3-1.2';
-import {
-  DirectoryApp,
-  DirectoryAppLaunchDetailsWeb,
-} from '/@/directory/directory';
-import { getSailManifest } from '/@/directory/directory';
+import { DirectoryApp } from '/@/directory/directory';
 import { FDC3Message, OpenData, TargetIdentifier } from '/@/types/FDC3Message';
+import { openApp } from '../lib/open';
 
 /**
  *
@@ -33,34 +29,7 @@ export const open = async (message: FDC3Message) => {
 
   if (allResults.length > 0) {
     const directoryEntry: DirectoryApp = allResults[0];
-    const start_url = (directoryEntry.details as DirectoryAppLaunchDetailsWeb)
-      .url;
-    const manifest = getSailManifest(directoryEntry);
-
-    let newView: View | void;
-
-    //if manifest is set to force a new window, then launch a new workspace
-    if (manifest.forceNewWindow && manifest.forceNewWindow === true) {
-      newView = await runtime.createView(start_url, {
-        directoryData: directoryEntry,
-      });
-    } else {
-      //else get target workspace
-      const sourceView = runtime.getView(message.source);
-      const work =
-        runtime.getWorkspace(message.source) ||
-        (sourceView && sourceView.parent);
-      newView =
-        work &&
-        (await work.createView(start_url, { directoryData: directoryEntry }));
-    }
-
-    //set provided context
-    if (newView && data.context) {
-      newView.setPendingContext(data.context, message.source);
-    }
-
-    return;
+    return await openApp(directoryEntry, message.source, data.context);
   }
   throw new Error(OpenError.AppNotFound);
 };
